requirements.txt
# ZeroMQ messaging library for client-server communication
pyzmq==25.1.2

# For potential HTTP requests to external APIs (e.g. TheMealDB)
requests==2.31.0

# Web scraping dependency for the server (uses Chrome WebDriver)
selenium==4.17.2

# Interactive command-line prompt (used in client)
questionary==2.0.1

# Terminal styling for clean output (used in client)
clint==0.5.1
